refactor: 회원가입 dto 구조 개선
