+++ modules/auth/src/main/java/kr/flint/auth/client/dto/KakaoUserInfo.java
@@ -1,25 +1,9 @@
 package kr.flint.auth.client.dto;
 
 public record KakaoUserInfo(
-        String providerUserId,
-        String email,
-        String nickname
+        String providerUserId
 ) {
     public static KakaoUserInfo from(KakaoUserResponse response) {
-        String email = null;
-        String nickname = null;
-
-        if (response.kakaoAccount() != null) {
-            email = response.kakaoAccount().email();
-            if (response.kakaoAccount().profile() != null) {
-                nickname = response.kakaoAccount().profile().nickname();
-            }
-        }
-
-        return new KakaoUserInfo(
-                String.valueOf(response.id()),
-                email,
-                nickname
-        );
+        return new KakaoUserInfo(String.valueOf(response.id()));
     }
 }
