+++ modules/auth/src/main/java/kr/flint/auth/client/dto/KakaoUserResponse.java
@@ -1,17 +1,5 @@
 package kr.flint.auth.client.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 public record KakaoUserResponse(
-        Long id,
-        @JsonProperty("kakao_account") KakaoAccount kakaoAccount
-) {
-    public record KakaoAccount(
-            String email,
-            KakaoProfile profile
-    ) {
-        public record KakaoProfile(
-                String nickname
-        ) {}
-    }
-}
+        Long id
+) {}
