+++ modules/auth/src/main/java/kr/flint/auth/dto/response/TempTokenPayload.java
@@ -1,0 +1,8 @@
+package kr.flint.auth.dto.response;
+
+import kr.flint.auth.domain.enums.AuthProvider;
+
+public record TempTokenPayload(
+        AuthProvider provider,
+        String providerUserId
+) {}
